// src/redux/car-installment/car-installment/initial-type.ts
export interface InitialType {
  installments_duration: number;
  loan_amount: number;
  refund_total: number;
  facility_amount: number;
  number_of_installment: number;
  facility_interest_percentage: number;
  facility_interest: number;
  marketing_percentage: number;
  check_installments_monthly: number;
  last_check: number;
  loan_advance: number;
  showFactor: boolean;
  leasing_id: string;
  brand_models: any[];
  brand: string;
  model: string;
  brandModelError: string;
  current_step: number;
  advertiser_id: string;
  ad_code: string;
  monthly_installment: number;
}